# api/internal/docs/generated/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TheTipTop API",
        "title": "TheTipTop",
        "contact": {},
        "version": "dev"
    },
    "host": "localhost",
    "paths": {
...
        "/employee": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Update a employee.",
                "operationId": "jwt.Auth =\u003e user.UpdateEmployee",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Newsletter",
                        "name": "newsletter",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "With the bearer started",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Password updated"
                    },
                    "400": {
                        "description": "Invalid email, password or token"
                    },
                    "404": {
                        "description": "Employee not found"
                    },
                    "409": {
                        "description": "Employee already validated"
                    },
                    "410": {
                        "description": "Token expired"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/employee/register": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Register a employee.",
                "operationId": "user.RegisterEmployee",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "default": "user-thetiptop@yopmail.com",
                        "description": "Email address",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Aa1@azetyuiop",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Employee created"
                    },
                    "400": {
                        "description": "Invalid email or password"
                    },
                    "409": {
                        "description": "Employee already exists"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/employee/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Get a employee by ID.",
                "operationId": "jwt.Auth =\u003e user.GetEmployee",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "With the bearer started",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee details"
                    },
                    "400": {
                        "description": "Invalid employee ID"
                    },
                    "404": {
                        "description": "Employee not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Delete a client by ID.",
                "operationId": "jwt.Auth =\u003e user.DeleteEmployee",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "With the bearer started",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Employee deleted"
                    },
                    "400": {
                        "description": "Invalid employee ID"
                    },
                    "404": {
                        "description": "Employee not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}